+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 增加部分服务的抽象实现。
   - com.dwarfeng.subgrade.impl.service.AbstractBatchCrudService。
   - com.dwarfeng.subgrade.impl.service.AbstractCrudService。
+  - com.dwarfeng.subgrade.impl.service.AbstractEntireLookupService。
 
 - 优化部分服务的实现的继承关系。
   - com.dwarfeng.subgrade.impl.service.CustomBatchCrudService。
@@ -15,6 +16,8 @@
   - com.dwarfeng.subgrade.impl.service.DaoOnlyCrudService。
   - com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService。
   - com.dwarfeng.subgrade.impl.service.GeneralCrudService。
+  - com.dwarfeng.subgrade.impl.service.GeneralEntireLookupService。
+  - com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService。
 
 - 优化部分接口及其实现的参数名称。
   - com.dwarfeng.subgrade.stack.service.BatchCrudService。